requirements.txt
mysql-connector-python
pandas
python-dotenv
openpyxl
mysql-connector
